% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal_places.R
\name{decimal_places}
\alias{decimal_places}
\title{Return the number of decimal places in a given value}
\usage{
decimal_places(x)
}
\arguments{
\item{x}{The value for which the number of decimal places will be returned.}
}
\value{
An integer representing the number of decimal places in x.
}
\description{
Return the number of decimal places in a given value
}
\examples{
decimal_places(4.56)
}
