% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{download_malert_zip}
\alias{download_malert_zip}
\title{Download Mosquito Alert zip archive (internal helper)}
\usage{
download_malert_zip(source, doi, destfile)
}
\arguments{
\item{source}{String. Source to download from. Options are "github" or "zenodo".}

\item{doi}{String. Zenodo doi if downloading from Zenodo.}

\item{destfile}{String. Path where the zip file should be saved.}
}
\value{
String. Path to the downloaded zip file.
}
\description{
Download Mosquito Alert zip archive (internal helper)
}
\keyword{internal}
