% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_malert_aggregates.R
\name{get_malert_aggregates}
\alias{get_malert_aggregates}
\title{Generates count by country of mosquito alert reports}
\usage{
get_malert_aggregates(
  aggregate_type,
  filter_year,
  file_path,
  country_code,
  file_layer
)
}
\arguments{
\item{aggregate_type}{String. type of aggragation you would like. Options are country or city}

\item{filter_year}{String. The year(s) you would like to filter for. Default is all. "2014,2"}

\item{file_path}{String. Link the the geopackage or shapefile you would like to use.}

\item{country_code}{String. type of aggragation you would like. Options are country or city}

\item{file_layer}{String. the layer of the shapefile/geopackage to access}
}
\value{
The aggregated data.
}
\description{
Generates count by country of mosquito alert reports
}
\examples{
malert_reports = get_malert_data(source = "github")
malert_reports
}
