% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_aggregated_time_series.R
\name{make_aggregated_time_series}
\alias{make_aggregated_time_series}
\title{Make complete time series aggregated by a given interval}
\usage{
make_aggregated_time_series(
  x,
  start = NA,
  end = NA,
  interval = "day",
  groups = c("trap_ID", "TigacellID")
)
}
\arguments{
\item{x}{A tibble containing exact datetimes for mosquito captures.}

\item{start}{A date or datetime object that will be used as the lower bound to filter records.}

\item{end}{A date or datetime object that will be used as the upper bound to filter records.}

\item{interval}{A string representing the interval over which data should be aggregated. Could be any string allowed by seq.Date.}

\item{groups}{A character vector representing the grouping variables to be included in the output.}
}
\value{
A tibble.
}
\description{
Make complete time series aggregated by a given interval
}
