% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lonlat_from_samplingcell_ids.R
\name{make_lonlat_from_samplingcell_ids}
\alias{make_lonlat_from_samplingcell_ids}
\title{Extract longitudes or latitudes from sampling cell IDs.}
\usage{
make_lonlat_from_samplingcell_ids(ids, type)
}
\arguments{
\item{ids}{A character vector of sampling cell IDs}

\item{type}{A string indicating whether longitude ("lon") or latitude ("lat") should be extracted.}
}
\value{
A numeric vector of longitude or latitude values.
}
\description{
Extract longitudes or latitudes from sampling cell IDs.
}
\examples{
make_lonlat_from_samplingcell_ids(ids=c("2.15_41.35", "2.10_41.20"), type="lon")
}
