% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mask.R
\name{make_mask}
\alias{make_mask}
\title{Mask a number by rounding it down to the nearest mask value. Used by 'make_sampling_cells' for transforming exact locations into masked sampling cells. This is basically the same as the round_down function but it returns a character instead of a numeric.}
\usage{
make_mask(x, mask = 0.05)
}
\arguments{
\item{x}{The number to be masked}

\item{mask}{The masking value.}
}
\value{
A character representing the result of the masking.
}
\description{
Mask a number by rounding it down to the nearest mask value. Used by 'make_sampling_cells' for transforming exact locations into masked sampling cells. This is basically the same as the round_down function but it returns a character instead of a numeric.
}
\examples{
make_mask(4.569, 0.05)
}
