% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_metadata_template.R
\name{make_metadata_template}
\alias{make_metadata_template}
\title{Make metadata tempplate ready for data portal}
\usage{
make_metadata_template(
  type = "Dataset",
  name = "",
  description = "",
  conditions_of_access = "",
  url = "",
  same_as = "",
  identifier = "",
  license = "",
  citation = "",
  measurement_technique = "",
  temporal_coverage = "",
  creator = "",
  variable_info,
  distribution
)
}
\value{
A json string.
}
\description{
Make metadata tempplate ready for data portal
}
