% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_samplingcell_ids.R
\name{make_samplingcell_ids}
\alias{make_samplingcell_ids}
\title{Creates standard sampling cell IDs by masking a set of longitude and latitude values.}
\usage{
make_samplingcell_ids(lon, lat, mask = 0.05)
}
\arguments{
\item{lon}{A vector of longitudes to be masked}

\item{lat}{A vector of latitudes to be masked}

\item{mask}{The masking value.}
}
\value{
A character vector of sampling cell IDs.
}
\description{
Creates standard sampling cell IDs by masking a set of longitude and latitude values.
}
\examples{
make_samplingcell_ids(lon=c(2.1686, 2.1032), lat=c(41.3874, 41.2098), 0.05)
}
