% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_covered_by.R
\name{not_covered_by}
\alias{not_covered_by}
\title{Simple predicate function for sf to filter all objects that are not covered by another one}
\usage{
not_covered_by(x, y)
}
\arguments{
\item{x}{Object to be filtered.}

\item{y}{Object used for filtering.}
}
\value{
Boolean.
}
\description{
Simple predicate function for sf to filter all objects that are not covered by another one
}
