% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{read_malert_json_RcppSimdJson}
\alias{read_malert_json_RcppSimdJson}
\title{Read a single Mosquito Alert JSON file using RcppSimdJson (internal helper)}
\usage{
read_malert_json_RcppSimdJson(file_path)
}
\arguments{
\item{file_path}{String. Path to the .json file to read.}
}
\value{
A tibble containing the flattened data for that file (year).
}
\description{
Read a single Mosquito Alert JSON file using RcppSimdJson (internal helper)
}
\keyword{internal}
