% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_down.R
\name{round_down}
\alias{round_down}
\title{Round a given number x downward to the nearest n.}
\usage{
round_down(x, n)
}
\arguments{
\item{x}{The number to be rounded.}

\item{n}{The number to be rounded.}
}
\value{
The numeric result of the downward rounding.
}
\description{
Round a given number x downward to the nearest n.
}
\examples{
round_down(4.569, 0.05)
}
