% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std2.R
\name{std2}
\alias{std2}
\title{Center and double-standardize a variable}
\usage{
std2(x)
}
\arguments{
\item{x}{A numeric vector to be centerd and double standardized.}
}
\value{
A numeric vector.
}
\description{
Center and double-standardize a variable
}
