% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_malert_data.R
\name{get_malert_data}
\alias{get_malert_data}
\title{Download Mosquito Alert report data from GitHub}
\usage{
get_malert_data(source = "github")
}
\arguments{
\item{source}{String. Source to download from. Currently only github works.}
}
\value{
A tibble.
}
\description{
Download Mosquito Alert report data from GitHub
}
\examples{
malert_reports = get_malert_data(source = "github")
malert_reports
}
