% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_malert_data.R
\name{get_malert_data}
\alias{get_malert_data}
\title{Download Mosquito Alert report data from GitHub}
\usage{
get_malert_data(source = "zenodo", doi = "10.5281/zenodo.597466")
}
\arguments{
\item{source}{String. Source to download from. Options are github or zenodo.}

\item{doi}{String. Zenodo doi if downloading from Zenodo. Default is the doi that will always point to the most recent version: 10.5281/zenodo.597466.}
}
\value{
A tibble.
}
\description{
Download Mosquito Alert report data from GitHub
}
\examples{
malert_reports = get_malert_data(source = "github")
malert_reports
}
