% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_senscape_data.R
\name{get_senscape_data}
\alias{get_senscape_data}
\title{Download data from Senscape server using http get request.}
\usage{
get_senscape_data(
  api_key,
  page_size = 10,
  start_datetime,
  end_datetime,
  deviceIds = NA
)
}
\arguments{
\item{api_key}{Path to Senscape API key.}

\item{page_size}{The number items per page. Defaults to 10.}

\item{start_datetime}{A datetime object that will be used as the lower bound to filter records.}

\item{end_datetime}{A datetime object that will be used as the upper bound to filter records.}

\item{deviceIds}{A character vector of all deviceIds from which data should be returned. If NA, then data will not be filtered by deviceId.}
}
\value{
A tibble.
}
\description{
Download data from Senscape server using http get request.
}
\examples{
new_smart_trap_data = get_senscape_data(api_key = ifelse(length(Sys.getenv('SENSCAPE_API_KEY'))>1, Sys.getenv('SENSCAPE_API_KEY'), keyring::key_get("SENSCAPE_API_KEY")), start_datetime = lubridate::as_datetime("2023-03-08"), end_datetime = lubridate::as_datetime("2023-03-09"), deviceIds = c("5f1076c998fda900151ff683", "5f1076c998fda900151ff683", "5f10762e98fda900151ff680", "5f10767c98fda900151ff681", "5f1076ae98fda900151ff682"))
new_smart_trap_data
}
