% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_senscape_devices.R
\name{get_senscape_devices}
\alias{get_senscape_devices}
\title{Download device information from Senscape server using http get request.}
\usage{
get_senscape_devices(api_key, page_size = 10)
}
\arguments{
\item{api_key}{Senscape API key.}

\item{page_size}{The number items per page. Defaults to 10.}
}
\value{
A tibble.
}
\description{
Download device information from Senscape server using http get request.
}
\examples{
my_devices = get_senscape_devices(api_key = Sys.getenv("SENSCAPE_API_KEY"))
my_devices
}
