% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_senscape_devices.R
\name{get_senscape_devices}
\alias{get_senscape_devices}
\title{Download device information from Senscape server using http get request.}
\usage{
get_senscape_devices(api_key)
}
\arguments{
\item{key_path}{Path to Senscape API key.}
}
\value{
A tibble.
}
\description{
Download device information from Senscape server using http get request.
}
\examples{
my_devices = get_senscape_devices(api_key = keyring::key_get("SENSCAPE_API_KEY"))
my_devices
}
